+++ src/main/java/org/mandas/docker/client/messages/mount/Driver.java
@@ -36,7 +36,6 @@ public interface Driver {
   @JsonProperty("Name")
   String name();
 
-  @Nullable
   @JsonProperty("Options")
   Map<String, String> options();