+++ src/main/java/org/mandas/docker/client/messages/swarm/DnsConfig.java
@@ -23,7 +23,6 @@ package org.mandas.docker.client.messages.swarm;
 import java.util.List;
 
 import org.immutables.value.Value.Immutable;
-import org.mandas.docker.Nullable;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -32,15 +31,12 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 @Immutable
 public interface DnsConfig {
 
-  @Nullable
   @JsonProperty("Nameservers")
   List<String> nameServers();
 
-  @Nullable
   @JsonProperty("Search")
   List<String> search();
 
-  @Nullable
   @JsonProperty("Options")
   List<String> options();