+++ src/main/java/org/mandas/docker/client/messages/swarm/NodeSpec.java
@@ -36,7 +36,6 @@ public interface NodeSpec {
   @JsonProperty("Name")
   String name();
 
-  @Nullable
   @JsonProperty("Labels")
   Map<String, String> labels();